move persisting functions to ln_storage
