+++ mm2src/coins/lightning/ln_utils.rs
@@ -2,16 +2,6 @@ use super::*;
 use crate::utxo::utxo_standard::UtxoStandardCoin;
 use common::mm_ctx::MmArc;
 use derive_more::Display;
-use lightning::routing::network_graph::NetworkGraph;
-use lightning::routing::scoring::Scorer;
-use lightning::util::ser::{Readable, Writeable};
-use secp256k1::PublicKey;
-use std::collections::HashMap;
-use std::fs::{File, OpenOptions};
-use std::io::{BufRead, BufReader, BufWriter, Read, Write};
-use std::net::{SocketAddr, ToSocketAddrs};
-use std::path::{Path, PathBuf};
-use std::str::FromStr;
 
 cfg_native! {
     use crate::utxo::rpc_clients::{electrum_script_hash, BestBlock as RpcBestBlock, ElectrumBlockHeader, ElectrumClient,
@@ -35,7 +25,8 @@ cfg_native! {
     use lightning::ln::channelmanager::{ChainParameters, ChannelManagerReadArgs, SimpleArcChannelManager};
     use lightning::ln::msgs::NetAddress;
     use lightning::ln::peer_handler::{IgnoringMessageHandler, MessageHandler, SimpleArcPeerManager};
-    use lightning::routing::network_graph::NetGraphMsgHandler;
+    use lightning::routing::network_graph::{NetworkGraph, NetGraphMsgHandler};
+    use lightning::routing::scoring::Scorer;
     use lightning::util::config::UserConfig;
     use lightning::util::ser::ReadableArgs;
     use lightning_background_processor::BackgroundProcessor;
@@ -45,9 +36,12 @@ cfg_native! {
     use lightning_persister::FilesystemPersister;
     use rand::RngCore;
     use rpc::v1::types::H256;
+    use secp256k1::PublicKey;
     use std::cmp::Ordering;
+    use std::collections::HashMap;
     use std::convert::TryInto;
-    use std::net::{IpAddr, Ipv4Addr};
+    use std::fs::File;
+    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
     use std::sync::{Arc, Mutex};
     use std::time::SystemTime;
     use tokio::net::TcpListener;
@@ -111,18 +105,6 @@ fn netaddress_from_ipaddr(addr: IpAddr, port: u16) -> Vec<NetAddress> {
     addresses
 }
 
-fn my_ln_data_dir(ctx: &MmArc, ticker: &str) -> PathBuf { ctx.dbdir().join("LIGHTNING").join(ticker) }
-
-pub fn nodes_data_path(ctx: &MmArc, ticker: &str) -> PathBuf { my_ln_data_dir(ctx, ticker).join("channel_nodes_data") }
-
-pub fn network_graph_path(ctx: &MmArc, ticker: &str) -> PathBuf { my_ln_data_dir(ctx, ticker).join("network_graph") }
-
-pub fn scorer_path(ctx: &MmArc, ticker: &str) -> PathBuf { my_ln_data_dir(ctx, ticker).join("scorer") }
-
-pub fn last_request_id_path(ctx: &MmArc, ticker: &str) -> PathBuf {
-    my_ln_data_dir(ctx, ticker).join("LAST_REQUEST_ID")
-}
-
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct LightningParams {
     // The listening port for the p2p LN node
@@ -175,7 +157,7 @@ pub async fn start_lightning(
     let broadcaster = Arc::new(platform_coin.clone());
 
     // Initialize Persist
-    let ln_data_dir = my_ln_data_dir(ctx, &ticker)
+    let ln_data_dir = ln_storage::my_ln_data_dir(ctx, &ticker)
         .as_path()
         .to_str()
         .ok_or("Data dir is a non-UTF-8 string")
@@ -317,8 +299,9 @@ pub async fn start_lightning(
 
     // Initialize the NetGraphMsgHandler. This is used for providing routes to send payments over
     let default_network_graph = NetworkGraph::new(genesis_block(network).header.block_hash());
-    let network_graph_path = network_graph_path(ctx, &ticker);
-    let network_graph = Arc::new(read_network_graph_from_file(&network_graph_path).unwrap_or(default_network_graph));
+    let network_graph_path = ln_storage::network_graph_path(ctx, &ticker);
+    let network_graph =
+        Arc::new(ln_storage::read_network_graph_from_file(&network_graph_path).unwrap_or(default_network_graph));
     let network_gossip = Arc::new(NetGraphMsgHandler::new(
         network_graph.clone(),
         None::<Arc<dyn Access + Send + Sync>>,
@@ -327,7 +310,7 @@ pub async fn start_lightning(
     let network_graph_persist = network_graph.clone();
     spawn(async move {
         loop {
-            if let Err(e) = save_network_graph_to_file(&network_graph_path, &network_graph_persist) {
+            if let Err(e) = ln_storage::save_network_graph_to_file(&network_graph_path, &network_graph_persist) {
                 log::warn!(
                     "Failed to persist network graph error: {}, please check disk space and permissions",
                     e
@@ -380,12 +363,14 @@ pub async fn start_lightning(
     ));
 
     // Initialize routing Scorer
-    let scorer_path = scorer_path(ctx, &ticker);
-    let scorer = Arc::new(Mutex::new(read_scorer_from_file(&scorer_path).unwrap_or_default()));
+    let scorer_path = ln_storage::scorer_path(ctx, &ticker);
+    let scorer = Arc::new(Mutex::new(
+        ln_storage::read_scorer_from_file(&scorer_path).unwrap_or_default(),
+    ));
     let scorer_persist = scorer.clone();
     spawn(async move {
         loop {
-            if let Err(e) = save_scorer_to_file(&scorer_path, &scorer_persist.lock().unwrap()) {
+            if let Err(e) = ln_storage::save_scorer_to_file(&scorer_path, &scorer_persist.lock().unwrap()) {
                 log::warn!(
                     "Failed to persist scorer error: {}, please check disk space and permissions",
                     e
@@ -872,97 +857,6 @@ async fn ln_node_announcement_loop(
     }
 }
 
-fn pubkey_and_addr_from_str(pubkey_str: &str, addr_str: &str) -> ConnectToNodeResult<(PublicKey, SocketAddr)> {
-    // TODO: support connection to onion addresses
-    let addr = addr_str
-        .to_socket_addrs()
-        .map(|mut r| r.next())
-        .map_to_mm(|e| ConnectToNodeError::ParseError(e.to_string()))?
-        .ok_or_else(|| ConnectToNodeError::ParseError(format!("Couldn't parse {} into a socket address", addr_str)))?;
-
-    let pubkey = PublicKey::from_str(pubkey_str).map_to_mm(|e| ConnectToNodeError::ParseError(e.to_string()))?;
-
-    Ok((pubkey, addr))
-}
-
-pub fn parse_node_info(node_pubkey_and_ip_addr: String) -> ConnectToNodeResult<(PublicKey, SocketAddr)> {
-    let mut pubkey_and_addr = node_pubkey_and_ip_addr.split('@');
-
-    let pubkey = pubkey_and_addr.next().ok_or_else(|| {
-        ConnectToNodeError::ParseError(format!(
-            "Incorrect node id format for {}. The format should be `pubkey@host:port`",
-            node_pubkey_and_ip_addr
-        ))
-    })?;
-
-    let node_addr_str = pubkey_and_addr.next().ok_or_else(|| {
-        ConnectToNodeError::ParseError(format!(
-            "Incorrect node id format for {}. The format should be `pubkey@host:port`",
-            node_pubkey_and_ip_addr
-        ))
-    })?;
-
-    let (pubkey, node_addr) = pubkey_and_addr_from_str(pubkey, node_addr_str)?;
-    Ok((pubkey, node_addr))
-}
-
-pub fn read_nodes_data_from_file(path: &Path) -> ConnectToNodeResult<HashMap<PublicKey, SocketAddr>> {
-    if !path.exists() {
-        return Ok(HashMap::new());
-    }
-    let mut nodes_data = HashMap::new();
-    let file = File::open(path).map_to_mm(|e| ConnectToNodeError::IOError(e.to_string()))?;
-    let reader = BufReader::new(file);
-    for line in reader.lines() {
-        let line = line.map_to_mm(|e| ConnectToNodeError::IOError(e.to_string()))?;
-        let (pubkey, socket_addr) = parse_node_info(line)?;
-        nodes_data.insert(pubkey, socket_addr);
-    }
-    Ok(nodes_data)
-}
-
-pub fn save_node_data_to_file(path: &Path, node_info: &str) -> ConnectToNodeResult<()> {
-    let mut file = OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(path)
-        .map_to_mm(|e| ConnectToNodeError::IOError(e.to_string()))?;
-    file.write_all(format!("{}\n", node_info).as_bytes())
-        .map_to_mm(|e| ConnectToNodeError::IOError(e.to_string()))
-}
-
-pub fn save_network_graph_to_file(path: &Path, network_graph: &NetworkGraph) -> EnableLightningResult<()> {
-    let file = OpenOptions::new()
-        .create(true)
-        .write(true)
-        .open(path)
-        .map_to_mm(|e| EnableLightningError::IOError(e.to_string()))?;
-    network_graph
-        .write(&mut BufWriter::new(file))
-        .map_to_mm(|e| EnableLightningError::IOError(e.to_string()))
-}
-
-pub fn read_network_graph_from_file(path: &Path) -> EnableLightningResult<NetworkGraph> {
-    let file = File::open(path).map_to_mm(|e| EnableLightningError::IOError(e.to_string()))?;
-    NetworkGraph::read(&mut BufReader::new(file)).map_to_mm(|e| EnableLightningError::IOError(e.to_string()))
-}
-
-pub fn save_scorer_to_file(path: &Path, scorer: &Scorer) -> EnableLightningResult<()> {
-    let file = OpenOptions::new()
-        .create(true)
-        .write(true)
-        .open(path)
-        .map_to_mm(|e| EnableLightningError::IOError(e.to_string()))?;
-    scorer
-        .write(&mut BufWriter::new(file))
-        .map_to_mm(|e| EnableLightningError::IOError(e.to_string()))
-}
-
-pub fn read_scorer_from_file(path: &Path) -> EnableLightningResult<Scorer> {
-    let file = File::open(path).map_to_mm(|e| EnableLightningError::IOError(e.to_string()))?;
-    Scorer::read(&mut BufReader::new(file)).map_to_mm(|e| EnableLightningError::IOError(e.to_string()))
-}
-
 #[derive(Display)]
 pub enum ConnectToNodeRes {
     #[display(fmt = "Already connected to node: {}@{}", _0, _1)]
@@ -1062,29 +956,4 @@ pub fn open_ln_channel(
     channel_manager
         .create_channel(node_pubkey, amount_in_sat, push_msat, events_id, Some(user_config))
         .map_to_mm(|e| OpenChannelError::FailureToOpenChannel(node_pubkey.to_string(), format!("{:?}", e)))
-}
-
-pub fn save_last_request_id_to_file(path: &Path, last_request_id: u64) -> OpenChannelResult<()> {
-    let mut file = OpenOptions::new()
-        .create(true)
-        .write(true)
-        .open(path)
-        .map_to_mm(|e| OpenChannelError::IOError(e.to_string()))?;
-    file.write_all(format!("{}", last_request_id).as_bytes())
-        .map_to_mm(|e| OpenChannelError::IOError(e.to_string()))
-}
-
-pub fn read_last_request_id_from_file(path: &Path) -> OpenChannelResult<u64> {
-    if !path.exists() {
-        return MmError::err(OpenChannelError::InvalidPath(format!(
-            "Path {} does not exist",
-            path.display()
-        )));
-    }
-    let mut file = File::open(path).map_to_mm(|e| OpenChannelError::IOError(e.to_string()))?;
-    let mut contents = String::new();
-    let _ = file.read_to_string(&mut contents);
-    contents
-        .parse::<u64>()
-        .map_to_mm(|e| OpenChannelError::IOError(e.to_string()))
 }
